services/collector/requirements.txt
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
